[rb] Support overriding default locator conversion
